+++ front-end/src/main.rs
@@ -4,13 +4,14 @@ use pasfmt_orchestrator::predule::*;
 
 pasfmt_config!(Config);
 
-fn main() {
+fn main() -> anyhow::Result<()> {
     let config = Config::create();
     stderrlog::new()
         .verbosity(config.log_level())
         .init()
         .unwrap();
 
-    let formatting_settings = config.get_config_object::<FormattingSettings>();
+    let formatting_settings = config.get_config_object::<FormattingSettings>()?;
     format_with_settings(formatting_settings, config);
+    Ok(())
 }